requirements.txt
Flask==2.0.2
lru-expiring-cache==1.1
python-dotenv==0.19.2
redis==4.0.2
requests==2.22.0